# docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080/api
  description: Generated server url
paths:
...
  /v1/user:
    get:
      tags:
      - user-controller
      operationId: profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      tags:
      - user-controller
      operationId: registration
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistrationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /v1/fx/master-list/economic-indicator/{countryCode}:
    get:
      tags:
      - master-list-controller
      operationId: economicIndicator
      parameters:
      - name: countryCode
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KeyValue"
...
components:
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID(Number)
          example: 999
        userId:
          type: string
          description: ユーザID(Cognito sub)
          example: 1aaaaaa1-2bb2-3cc3-4dd4-5eeeeeeeeee5
        emailAddress:
          type: string
          description: email
          example: account@domain.com
        nickName:
          type: string
          description: nickname
          example: Mr. Consideration
        approved:
          type: boolean
          description: 承認フラグ
        approvedAt:
          type: string
          format: date-time
          description: 承認日時
          example: 2026-01-23T12:34:56+09:00
        admin:
          type: boolean
          description: 管理者フラグ
        blocked:
          type: boolean
          description: auth0 blocked
        createdAt:
          type: string
          format: date-time
          description: 新規登録日時
          example: 2026-01-23T12:34:56+09:00
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
          example: 2026-01-23T12:34:56+09:00
      required:
      - admin
      - approved
      - blocked
      - createdAt
      - emailAddress
      - id
      - nickName
      - userId
    UserResponse:
      type: object
      properties:
        returnCode:
          type: integer
          format: int32
          description: 処理結果
        message:
          type: string
          description: message
        user:
          $ref: "#/components/schemas/UserDto"
          description: ユーザー情報
      required:
      - returnCode
      - user
...
    ApiResponse:
      type: object
      properties:
        returnCode:
          type: integer
          format: int32
          description: 処理結果
        message:
          type: string
          description: message
      required:
      - returnCode
    UserRegistrationRequest:
      type: object
      properties:
        nickName:
          type: string
          description: ニックネーム
          example: Mr. Consideration
          maxLength: 50
          minLength: 0
      required:
      - nickName
...
    KeyValue:
      type: object
      properties:
        key:
          type: string
        value:
          type: string